feat(backend): Refactor backend logic to support multiple endpoints
